requirements.txt
aiohttp==3.7.4.post0
amqp==5.0.6
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
billiard==3.6.4.0
celery==5.1.1
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.3
django-celery-beat==2.2.0
django-celery-results==2.1.0
django-cors-headers==3.7.0
django-timezone-field==4.1.2
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
kombu==5.1.0
multidict==5.1.0
prompt-toolkit==3.0.19

psycopg2-binary==2.8.5
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
yarl==1.6.3
sentry-sdk==0.10.2
PyJWT==2.1.0
djangorestframework-simplejwt==4.8.0
# djangorestframework-simplejwt
django-extensions
django-import-export
django-admin-rangefilter
django-queryset-csv
google-auth
